Introduce a wrapper around hyperlight-wasm
This is a really simplistic wrapper around the hyperlight WASM
library. It will build a static library with an C/C++ ABI
compatible interface.

It's simplistic in this sense:

1. No thoughts were given to how the C/C++ interface should really
   look like, so the only consideration that was given is
   simplisity of use (e.g. no considerations were given to how
   error-prone it is or whether it allows to use hyperlight
   efficiently)
2. Error handling is basically missing, so it's quite fragile
3. It does not expost knobs to enable configuration and tunning
   of hyperlight to the specific use cases.

IOW, it's just for the proof-of-concept and should not really be
used as-is without modifications and giving much more thoughts to
it.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>