Scripts to build Envoy and WASM filter
The WASM filter I use here is very simplistic. All it
does is takes the data passed to in into the onData
function and call external function write to send that
data along.

I didn't figure out the structure of WASM executables,
how to build them and what should they expose to work
with Hyperlight and WAMR inside it. So I just replicate
the build setup used in https://github.com/deislabs/hyperlight-wasm.
It may take some time to build the Docker image with the
SDK, but we will have to get by for now.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>