Bump version of Envoy and start using release versions
While debugging versions are extremelly useful, we do want to understand
performance of the release build, as it's what will likely be used in
real-life scenarios (even though we are a bit far from that at the moment).

For debugging build all it takes is to change the bazel build flag in the
prepare.sh script, so using release version by default wouldn't be that
big of an obstacle.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>