Echo filter implementation using proxy-wasm
The intention is to be able to compare proxy-wasm performane
with hyperlight performance. That is not going to be an apples
to apples comparision, because proxy-wasm specification interface
is somewhat different from what I have implemented so far for
hyperlight. That being said, I don't think that it renders the
compariosion copletely meaningless.

I want to understand what kind of performance does proxy-wasm
extension is currently capable of, even if the details of the
implementation of such an extension are different from hyperlight.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>