Drastically improve performance by reducing wasm heap size
I found out that in WAMR the cost of wasm_runtime_instantiate
depends on the size of the heap used in wasm. That's because
they explicitly memset the content of the heap after allocation.

Additionally, hyperlight-wasm guest currently calls
wasm_runtime_instantiate during each guest call. So in our
Hyperlight/Envoy PoC integration we pay that cost on every onData
callback, which ends up being quite expensive.

This change reduces the size of the heap we use in PoC to a bare
minimum. It's not a solution to the problem (e.g. we definitely
can do better than that by just moving wasm_runtime_instantiate
out of the data path to the initialization stage), but it helps
to have a better performance for the demo while the solution
is not yet ready.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>